Issue #5: Logon
Adds logon with auth0.

Creates a <logon-required> tag component.

The contents of this element are shown if logged-in,
else a logon form.

Adding AuthHttp provider in app.ts broke tests in app.spec.ts.
I fought this for a while before commenting it out.

Fought InAppBrowser not working with Auth0Lock,
not sure how, but got it working on android device.

Adds package.json for intl.

Closes #5.