Issue #66: Refresh Pages on Pop if Data May Have Changed
In the places where we push/pop pages onto the Nav stack,
if the data has changed, publish an event.
The pages behind the new page will subscribe to this event,
and refresh their data when it is fired.

This was changed for the new-dose-medication page, which
fires an event that refreshes the med-list page.

This was changed for new-amigo-request, which fires an event
that refreshes the amigos and requests pages.