Issue #70: Fix Logon bug.
Quick bug fix, not ideal solution.
Since the feed page is the most likely page this will occur on,
don't issue the server calls until after the user logs in.
"user:authenticated" event will trigger the reload attempt.

Also removes the caching of the current user.
If this fails, we are stuck with bad user data.
Also since this contains the dose information, we need it to refresh.

Changes slogan text color now that theme colors changed.